requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.3.1
billiard==3.6.3.0
black==20.8b1
celery==5.0.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
conda==4.3.16
distlib==0.3.1
Django==3.1.4
django-crispy-forms==1.10.0
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2020.8.13
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
pyflakes==2.2.0
pytz==2020.4
redis==3.5.3
regex==2020.11.13
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
